Background materials/03-Labs/Lab-03-Multi-Agent/starter/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
chromadb>=0.4.0
